[core][sparse][pruning] cuSPARSELt Kernels and ops.
Summary:

This is a duplicate PR of 102133, which was reverted because it was
failing internal tests.

It seems like that internal builds did not like my guard to check if
cuSPARSELt was available or not.

The linked Diff includes updates to TARGETS and
oversource_aten_gen_defsf.bzl, to replace

@AT_CUSPARSELT_ENABLED@ with the appropriate 0/1 value.

This PR also updates cusparselt to get the current CUDA stream, rather
than passing in nullptr, which resolves some problems when calling
torch.compile

Test Plan:

python test/test_sparse_semi_structured

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: d29b218fad2462a567452027972bdd593b6d9fa8
Pull Request resolved: https://github.com/pytorch/pytorch/pull/107176