[inductor] Separate to_{dtype,device} from lowering to avoid copying
These lowerings must copy even when they are no-ops in order to preserve
correctness in the presense of mutations. However, `to_dtype` and `to_device`
are also used in various lowerings as a helper function where it is okay to alias.

So, I've split these into two functions and allow the helper functions to alias
which saves some unnecessary copies.

ghstack-source-id: c563be540361e8f780c179363796ef60fa7e9579
Pull Request resolved: https://github.com/pytorch/pytorch/pull/107640