Ensure that FakeTensor.data_ptr() throws
We (for some reason) have two paths for data_ptr accesses: one for
templated data_ptr and one for untemplated data_ptr.

The templated data_ptr path raises when the input is a FakeTensor. This
PR changes it so that if a new `data_throws_if_uninitialized_` field is set
on TensorImpl, then untemplated data_ptr access (which goes through a
helper function called TensorImpl::data()) will throw on FakeTensor.

Afaict we cannot unconditionally throw on untemplated data_ptr access
if the storage is uninitialized because (according to the docs) caffe2
relies on that, plus I don't want to regress perf for the regular Tensor
case too much.

Test Plan:
- some new tests

ghstack-source-id: 5ea9f329d18d2c890a1a99eeae04a3ef501c9171
Pull Request resolved: https://github.com/pytorch/pytorch/pull/107726