Add torch::Library::requesta_pyimport
We want to be able to define custom ops in Python and C++: a user may
put things like meta kernels in Python but the backend kernels in C++.

When someone is using a custom op from Python, we want to ensure that
both the C++ shared library and the Python kernels are loaded to avoid
partially-initialized custom ops.
When someone is using a custom op from C++, we assume they just want
to use the C++ kernels.

The design we use to prevent partially-initialized custom ops in
Python is that a C++ torch::Library object can import a python module
by calling `request_pyimport`, which will import the module if Python
is available. This ensures that when the C++ TORCH_LIBRARY block gets
loaded, the Python module will also be imported.

Implementation details:
- Due to the libtorch-libtorch_python split, we have a PyImportsHandler
indirection. PyImportsHandler has a method that imports a module in
Python.
- When `import torch` happens, we initialize the PyImportsHandler
- If a user wants to load a shared library with custom ops that calls
`request_pyimport`, they must do so AFTER `import torch` happens. If
they
do it before, we'll raise an error on `import torch`.
- The PyImportsHandler is protected by a mutex: TORCH_LIBRARY
static initialization may be multi-threaded.

Test Plan:
- new test

ghstack-source-id: d3c6c936075055f1e5eca331b280c3e14bef0afa
Pull Request resolved: https://github.com/pytorch/pytorch/pull/108771