Deprecate torch.cross default behaviour
Long overdue this one. We may be able to change it in a few years :hopeful:.

**BC-breaking note**

This PR deprecates `torch.cross`'s default dim in favor of
`torch.linalg.cross`.
A upgrade guide is added to the documentation for `torch.cross`.

Note this PR DOES NOT remove `torch.cross`.

Fixes https://github.com/pytorch/pytorch/issues/108664

ghstack-source-id: 3e7bae968f74a854c03e886d18597e6817d7cb41
Pull Request resolved: https://github.com/pytorch/pytorch/pull/108760