[inductor] Fix triton compiler error in multilayer any
Fixes #109196

When we have a split reduction and the tensor is not an even multiple of the split size,
we use `ops.masked` to pad to an even multiple. In the case here we generated:
```python
tmp5 = tl.where(mask, tmp4, 0)
```

which implicitly promotes our boolean value to `int32`. The fix is to give the default
value the same dtype as `result`.

ghstack-source-id: d3294c05eb8f46e10eebbcec05299819b4c8dad1
Pull Request resolved: https://github.com/pytorch/pytorch/pull/109325