Fix max/min.reduction_with_dim opinfo test for bool tensors
The tests were failing because the input tensors were getting
incorrectly promoted to int64, due to `transform_args` incorrectly
considering the `dim` argument when determining the type to promote to.

It doesn't seem like type promotion is necessary in general for max and
min, so I've switched them to `type_promotion_kind=None`.

ghstack-source-id: a12fb2a3362f18b977a0e12f5578abb58a55bbf0
Pull Request resolved: https://github.com/pytorch/pytorch/pull/109264